003_loading_data_from_www_page_to_bq/requirements.txt
functions-framework==3.* # Cloud Functions
google-cloud-bigquery # Cloud Functions
google-cloud-storage # Cloud Functions

Flask # local
Jinja2  # local
google-auth  # local


# pip install -r requirements.txt